Rename core scripts to remove 'new'
